Assignment added4

+++ src/main/java/com/driver/SavingsAccount.java
@@ -33,7 +33,7 @@ public class SavingsAccount extends BankAccount{
     public double getCompoundInterest(int times, int years){
         // Return the final amount considering that bank gives compound interest on current amount given times per year
         double balance = getBalance();
-        double power = times*years;
+        int power = times*years;
         double ci = balance * (Math.pow((1 + rate / times), power));
         return ci;
     }
